Update patch set 9
Patch Set 9:

> Patch Set 8: Code-Review+2
> 
> Thanks. Verified that the test vector mismatch is gone with PS8. Do you know what causes the mismatch?

Hi Yunqing,

It is seen that constant value of 2^(bd + FILTERBITS -1) is added to the horizontal filter output. As per our understanding this constant is added to make the output of horizontal filter non-negative. Adding a constant before horizontal filtering and subtracting the same after vertical filtering is redundant was our thought and we removed the same. But in wiener filtering the most negative horizontal output can be -90 x 255 for 8 bit case whereas the constant added i.e., 2^(bd + FILTERBITS -1) is 16,384 which is lesser than 90 x 255 = 22950. Thus the output of horizontal filter can be negative and it is later clipped to [0, 8191] before vertical filtering. Hence, adding a constant before horizontal filtering and subtracting the same after vertical filtering is not redundant. Thus we have added it back. 

Thanks and Regards
Satish

Patch-set: 9
Reviewer: Gerrit User 10122 <10122@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>