Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Hi Logan, this patch is stuck at a compiler error related to tensorflow include directory
> The error msg is in: https://build.aomedia.org/jenkins/job/libaom__compile_experiments/arch=generic-gnu,shard=9/17147/console
> 
> 14:47:45 CMake Error at CMakeLists.txt:220 (message):
> 14:47:45   TENSORFLOW_INCLUDE_DIR does not refer to an existing directory
> 
> I am curious if you have met similar errors before? I checked CMakeLists.txt before and after this CL, nothing related to TENSORFLOW_INCLUDE_DIR is changed. It would be great if you can take a look

Hey Yue.

TENSORFLOW_INCLUDE_DIR doesn't have any default path, but if -DCONFIG_TENSORFLOW=1, then it needs to be fed a valid path to the root of a tensorflow repository. 

It looks like the grep statement is searching for experiments AV[12]*. I don't think Jenkins has a tf directory, so we would have to disable the check for this experiment in this patch. Perhaps we could rename the description to "AV2 TF experiment flag" or something along those lines? I can check with James/Tom to see how we could address this issue so we can develop a plan to handle TF in Jenkins in the future.

Patch-set: 2
Reviewer: Gerrit User 11494 <11494@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>