Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> Deb, look like sf->inter_mode_rd_model_estimation = 1; is set at speed 0. So, with this CL, in multi-thread case, sf->inter_mode_rd_model_estimation = 2; will be used at speed 0. Maybe we could run some quality test?

Yes quality tests are already done.
Between type 1 and 2, we can have two settings: in one case type 2 is a little slower than type 1, but there is an improvement of -0.07 on lowres, -0.02 on midres. In an alternative setting type 2 is generally faster than type 1 in speeds 1-4 and similar in speed 0, but there is a loss of +0.015% on lowres and +0.08% on midres. Current values are tuned for the second case since I wanted type 2 to be faster.
if we wanted to change to the first case, that is a simple one-line change. Let me know what you think about whether we should do that. 
Note that in either case it will be faster than setting to 0 in the multi-thread case.

Patch-set: 3
Reviewer: Gerrit User 5189 <5189@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>