Update patch set 1
Patch Set 1:

Yunqing's suggestion from offline discussion:
"This would help in some cases, but it wouldn't help if incorrect distance < lag-in-frames. If we want to use the order hints in the encoder side, we can add a "real" order hint that is represented as an unsigned int in encoder side only. "

Based on that, holding onto this patch. Will wait for Yunqing to see if the above suggestion can be implemented instead.

Patch-set: 1
Reviewer: Gerrit User 5290 <5290@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>