Update patch set 1
Patch Set 1: Code-Review+2

Is there a list of these that are causing a problem? I think anything in the library is already prefixed w/av1 or aom, so you're just going to get aom_av1 or aom_aom as the outputs here. In the end it's harmless, so LGTM if it's fixing something or making life easier.

Patch-set: 1
Reviewer: Gerrit User 5215 <5215@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>
Label: Code-Review=+2