Update patch set 1
Patch Set 1: Code-Review+2

Yaowu: I guess you want to only allow global (non-static) functions to use the av1_/aom_ prefix?

For the aom_ prefix, which I believe is intended for the functions in the public API, this is a good change.

For the av1_ prefix, it seems harmless to allow some static functions to use the av1_ prefix. On the other hand, it is a good idea to be able to tell by a function's name whether it is a global or static function.

Patch-set: 1
Reviewer: Gerrit User 9545 <9545@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>
Label: Code-Review=+2