Only search compound modes with a good ref frame.
A reference frame is good if, among the single reference modes,
it was either the best frame or second best.

STATS_CHANGED
Lowres and midres test set indicate ~0.05 BD-rate drop for PSNR.
Speed-up ranges from 1% to 15% and is highly dependent on
target bitrate / speed setting. Biggest gains are when bitrate is
high and --cpu-used=0.

Change-Id: Ie290b721f5d952abdf61112c13fcc230a65cca02