Update patch set 1
Patch Set 1: Code-Review+2

We talked about this test on the vpx side, it's not really meaningful for one-pass non-realtime modes. It's not something that we've ever really focused on. The range is far too large and the number of frames from a single clip aren't enough to give a good signal.

Patch-set: 1
Reviewer: Gerrit User 5020 <5020@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>
Label: Code-Review=+2