Adjust gf_boost based on temporal stats
A small coding gain of about -0.04%. Hopefully will improve
with a better model from stats to boost.

Compared to no tpl_model or delta-q the gains are:
lowres (33 frames, cpu-used 0): -0.292%
midres (33 frames, cpu-used 0): -0.706%

Includes some refactoring in the process, including moving
the tpl stats processing for the frame to before the base_qindex
for the frame is decided.

STATS_CHANGED

Change-Id: Ia7daf6c3b4c6a2d7337c9a8bfabc121f70e4c5eb