Update patch set 4
Patch Set 4:

To answer a question asked offline:
I checked the patch on some more test sets that were NOT part of training/validation of the ML model:
- JVET 480p clips
- objective-1-fast set
- hdres set

And the performance was neutral on all of these clips. So the ML model seems generic enough to target clips that specifically benefit from disabling Alt-ref frames.

Patch-set: 4
CC: Gerrit User 5022 <5022@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>
CC: Gerrit User 5187 <5187@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>