Fix MV out of range bug
While encoding ultra high resolution(8k) video, the MV can go out of
range. The MVs obtained from single refernece frame modes are used
in compound reference modes. We need to check if the single reference MV
is valid, and also need to clamp it if it is out of range. Otherwise,
out of range MV is created.

This patch fixed the above bug.
BUG=aomedia:2408
This would potentially cause STATS_CHANGED.

Change-Id: Ibbaa33dc850b86563ab8e5448e4dec97e73e8371