Update rdcost after the change of block size
This is a preparation for adding tune=ssim mode. In tuning for SSIM
mode rdmult will be different in different blocks. So, we need to
recalculate the rdcost using the accumulated rates and distortions
whenever the block size changes in the recursive partition searching
stage.

STATS_CHANGED:
         avg_psnr  ovr_psnr  ssim
low_res  0.005     0.001     -0.025
mid_res  pending
hd_res   pending

Change-Id: Ica461b25ca85748bf4e9be59dc6bf5d8d310c809