av1_dec_fuzzer: get thread count from 1st byte of frame header
In most of the cases, 1st byte of the file header is 'D'.
So using first byte results in same thread count in most cases.
Using 33rd (offset 32, 1st byte of frame header) byte in the data
(one of the bytes that signals size of the frame) will help in
testing for different thread configurations.

Change-Id: I4494cad8e15d5c81920448904487ac24f7171e88