Introduce CNN-based intraframe partition on lowres
This commit adds a speed feature to use CNN to speed-up partition
search on intraframe. Currently this speed feature is only on
for lowres videos on speed 1, and only split/no_split decision is
made.

The CNN model operates on units of 64X64 block. It takes in 65X65
luma pixels (one row above and one colum left for the padding) and
log q_index as input features.

The model has about 8.5k float parameters, or about 34 kB in size.
Inference takes about 300ms per prediction. This translates to
about 1~2% overhead on easy videos when qp is high, but only
0.1~0.2% when qp is low.

Performance:
         | AVG_PSNR | OVR_PSNR |   SSIM   |  AVG_SPD  |  OVR_SPD
 KF-only |  0.117%  |  0.117%  |  0.042%  |  29.759%  |  35.672%
 33-frms |  0.046%  |  0.052%  | -0.006%  |   2.936%  |   2.582%

STATS_CHANGED

Change-Id: I9b5ee2f9a6f5dab61eab6dd108f4ce8bf3496f7a