Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > what's the behavior of this file in a release build prior to this patch?
> > 
> > It fails in get_tile_buffer() at av1/decode/decodeframe.c:2609:
> > 
> 
> Thanks, it sounds like it isn't a candidate for back port if we're able to fail gracefully.

The behavior depends on the test case. I only know what happens with the reproducer test case of this bug.

If you are considering back port, the complete fix should also include
https://aomedia-review.googlesource.com/c/aom/+/89963

I split the complete fix into two CLs because the check in the other CL is more complicated and I don't want that to block this CL.

Patch-set: 1