Recode loops for both superres and fullres.
This combined recode loop is used in SUPERRES_AUTO mode to find the
better out of the two as per RDCOST.

It is enabled for VBR and CQ modes, for video (non-image) coding.

Note that we need to save / restore the following extra info between encodes:
- Deblocking filter params, as previous values of
the params are used when searching for these params.
- cdef info for the same reason.
- The current buffer, as it may have been filtered
- RATE_CONTROL object inside 'cpi', as some updates there are meant to
be used for encoding the next frame

Also, for frames where superres is enabled, we force denom to be at
least 9 in superres pass, so that both fullres and superres are tried
in recode loop.

BDRate (PSNR) improvement for hdres set at low target bitrates:

(i) SUPERRES_AUTO mode (before vs after):
Overall: -1.79%
- station2_1080p25.y4m      -13.974
- stockholm_ter_720p60.y4m  -6.350
- mobcal_720p50.y4m         -6.182
- old_town_cross_720p50.y4m -5.631
- city_720p30.y4m           -4.297
- johnny_720p60.y4m         -4.089
- rush_hour_1080p25.y4m     -4.042
- shields_720p50.y4m        -3.384

(ii) fullres vs SUPERRES_AUTO (after):
Overall: -0.187
- shields_720p50.y4m:          -3.117
- factory_1080p30.y4m:         -1.640
- cyclists_720p30.y4m:         -1.564
- mobcal_720p50.y4m:           -1.413
- sheriff_720p30.y4m:          -1.311
- basketballdrive_1080p50.y4m: -0.739
- stockholm_ter_720p60.y4m:    -0.723

STATS_CHANGED for SUPERRES_AUTO mode

Change-Id: I4dd984f35f68d80409d1f53083eda00e257c1706