[Test] Recode loop: Try both superres and fullres.
Need to save / restore the following between encodes:
- Deblocking filter params, as previous values of
the params are used when searching for these params.
- cdef info for the same reason.
- The current buffer
- RATE_CONTROL object inside 'cpi', as some updates there are meant to
be used for encoding the next frame

Also, for frames where superres is enabled, we force denom to be at
least 9 in superres pass, so that both fullres and superres are tried
in recode loop.

STATS_CHANGED

Change-Id: I0412878c9b17c428d3748e491fc4b89b23a88c4b