Add a precision reduction step after nn_predict
The precision reduction step is a temporary fix to reduce the
probability of mismatches between C and SIMD implementations
of av1_nn_predict().

The patch fixes the issue in bug 2415, but there is no
guarantee that mismatch will not ever happen, since the
error is often larger than the reduced precision.

STATS_CHANGED

BUG=aomedia:2415

Change-Id: I70298e0e35abfe86cb65ad12b7ee506f9b736e74