Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > Is this similar to one of your vp9 changes?
> 
> I vaguely recall that we ended up reverting this change in VP9 since it caused a few clips performance drop in yt regression test. You might want to keep an eye on the regression test too this time.

As I understand correctly you are referring to the the following fix in VP9 which was related to scene change detection. (https://chromium-review.googlesource.com/c/webm/libvpx/+/1446343). However, this fix is in calculation of key frame boost. We have not submitted any patch in VP9 similar to this.

We found this issue in the code while trying to use undamped adjustment for updating the rate correction factors. This experiment is being tried in order to reduce the encode time.

Patch-set: 2
Reviewer: Gerrit User 8893 <8893@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>