Fix double clamping bug in 2 pass rate control.
In the two pass rate allocation the VBR min and max rates
are applied through normalization of the per frame error scores.

Hence when the bits for an ARF/GF group are calculated,
the contributions from each frame are already clamped to the
required VBR range.

After the total number of bits for the group (based on the
frames within that group) have been allocated, a proportion
of the bits from each normal frame are then diverted to boost
the ARF/GFs.

Prior to this fix, the rate for each frame was clamped again
after the redistribution to the ARF/GF which can fundamentally
break the rate control if two_pass_vbrmin_section is non 0.

For example, the default VBR settings used in the lowres test run
have VBRmin and VBRmax section values of 0 and 2000. This works fine
because the min value is 0. However, if the min value is changed to 50
then the rate control breaks down completely and there is a huge hit to
metrics.

For example in one of my test runs I saw.

Av PSNR,     Ov PSNR,    SSIM,         PSNR HVS
+25.65%     +63.66%   +38.36%    +29.4%

In fact for overall PSNR AV1 run drops below VP9.

After the patch the metrics impact is minimal.

Av PSNR,     Ov PSNR,    SSIM,         PSNR HVS
-0.0003       0.0003,      -0.0002      -0.0002

Using VBR defaults of 50 min and 150 max in my YT rates test set the
impact is also minimal.

STATS_CHANGED
Change-Id: I5936880706d2376e3cb03d241ace5b9eca9de99b