Explain why the code in get_txb_ctx is correct.
get_txb_ctx() calculates top and left differently from the spec, using
bitwise OR instead of the Max() function. Add a comment to justify this
method.

Background info:

This code was added in commit 4ab9a5dcd443fd5cc347fcea03e685b3644364cf
(reverted because of a bug in the dc_sign_contexts table):
    https://aomedia-review.googlesource.com/c/aom/+/39524
and commit ed629f8370e3a5bd9c3c1c157ca36c0f0cfa495e (second attempt):
    https://aomedia-review.googlesource.com/c/aom/+/39902
But the use of bitwise OR was not explained. I am not sure if this is
a well-known trick in certain circles, but it was not obvious to me
why it is equivalent to the spec.

Change-Id: Ia911065e2efd31157bbcabdf9b9d335a5cbec531