Update patch set 2
Patch Set 2:

Hi Hui Su,
There seems to be a quality issue with this commit. As per AWCY run below, this commit has a bd rate drop of ~2.01%
https://beta.arewecompressedyet.com/?job=rdmult_tweak_parent%402019-07-18T06%3A57%3A51.368Z&job=rdmult_tweak%402019-07-18T06%3A58%3A49.910Z
Commit message highlights about lowering rdmult, but as per the change, rdmult is multiplied by 13/4 which could be the reason.
Though we are not entirely sure, multiplication by 13/14 is possibly intended.
Can you check the same.
Regards,
Ranjit

Patch-set: 2
CC: Gerrit User 5185 <5185@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>
CC: Gerrit User 6245 <6245@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>