AV1 levels: change max bitrate constraint
The spec does not spicify clearly how the max bitrate constraint should be
applied. Before this patch, we checked the bitrate of any 1-second window.
With this patch, we check the average bitrate across the whole sequence
instead. The bitrate spikes can be controlled by the (per frame) minimum
compression ratio.

BUG=aomedia:2332

Change-Id: I7f6bd8abf368074b72c83da9a0fcd67a82c0ac98