Fix condition to try ext partitions
The extended partition should only be used when no more than half of the
current block is outside of the available image. Currently, the codebase
tests this condition with partition_none_allowed, which can be modified
by various speed features. Instead, the condition should be tested with
has_rows && has_cols.

Change-Id: I2e11730d586b9e746537ed3411afd908e08f7915