Update patch set 3
Patch Set 3:

In general tightening up rate control tends to give some metrics loss. Also in general a  tendency to undershoot hurts metrics less than overshoot. This is because if we overshoot early on we may run out of bits in the last few frames and this can result if a sharp dip in quality near the end as we target a lower and lower rate per frame. In contrast undershoot generally just results in gradual improvement in quality, so there is still variation from beginning to end but the worst case behavior is not as bad as the other way around.

Patch-set: 3