Update patch set 2
Patch Set 2:

This change seems correct and is in line with what is in VP9. I suspect the small regression will resolve with a little tuning of the key frame boost calculation and if other changes to that calculation are imported from VP9.

Patch-set: 2
Reviewer: Gerrit User 5610 <5610@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>