Some fixes for usage of externl frame flags.
For the case where the flags are passed in externally:
allow for the condition that LAST need not be used as
a reference.

Also use the correct ref_idx for LAST when setting the
efresh mask, when ext_refresh_frame_flags_pending is set.

For the issue of LAST as a reference: added an example for
3 temporal layers, where the second top enhancement layer
only predicts/references GF.

Change-Id: Ic5f7cef5032e214ae0df4c61069392a425e40f99