Extend encoder control for refreshing all buffer slots.
Extend the existing control to allow for refreshing any of the
allowed 8 buffer slots (current AOM_EFLAGS_UPD_ don't allow for
updating LASTS/LAST3, for example). So this change removes the
need to set the AOM_EFLAG_NO_UPD_flags.

This allows for more flexibility for RTC applications to
configure SVC structures.

Change-Id: I38fc957943d968636df9434071da82db335a2c30