Update patch set 4
Patch Set 4: Code-Review+2

This change covers some of the same ground as two changes to libvpx dated from Q4 2016 that changed the basis of boost calculation for key frames.

Change-Id: Iaa15ac5a58d2be89181100d95cef6a8dc4b12d0d
and 

Change-Id: I0fb2344d0f78b5e95bb899dfad12b0ca84034b2c

The latter also removes the average decay accumulator.

I think it may be worth following up on the other changes made for libvpx including maximum boost values for key frames which I have also tentatively trialed over the last couple of weeks in AOM. For now though I think this change is reasonable as a first step and in line with changes found to be beneficial in libvpx. We may well need to make further adjustments though in response to other changes and bug fixes such as:-

Change-Id: I51de19937e20a24dd1a22106c506a0308ca41cf1

Which fixes a bug in the calculation of KF boost but may well have acted to reduce average boost.

Patch-set: 4
Reviewer: Gerrit User 5610 <5610@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>
Label: Code-Review=+2