Revert "Scene-cut detection: avoid very short KF intervals"
This reverts commit ca3651934f9ec2bf70cd855b5cddce9124ef317e.

This patch causes a regression on some clips in the hd set though the exact
behavior depends a lot on other patches that impact the KF and arf boost.

At the current head this change causes a large regression for mob_cal but helps
for shields and stockholm. However, tested at the point in the tree it was
submitted it was hurting all three of these. This change may still sense
ultimately but is reverted for now until other changes to tune kf and arf boost
are complete.

STATS_CHANGED:

Change-Id: I460b2df291a6d67700fbb75da65adc0c40c16d63