Update patch set 1
Patch Set 1:

I think using the limit to signal that we want still picture was a weird way to do this.   Better to positively state - create a still picture frame with set 1.   If we do that and give it more than 1 frame it should either fail or create a sequence of still frames.

Patch-set: 1
Reviewer: Gerrit User 5010 <5010@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>