Update patch set 3
Patch Set 3: Code-Review-1

Whilst adjustment to the minQ allowed for kfs may well be needed and justified I am not sure this is the right place to do it.

The factor here looks like it was intended to offer a specific reduction in minQ for very small image formats where. These formats tend to have higher average spatial (intra complexity / error) compared to the average best/inter coding error and this may be the justification for allowing a larger q delta on the key frames. Really though this factor as implemented here will only affect formats such as 240P which are not heavily represented in our routine testing.

It IS quite likely that we may need to adjust the calculation of the allowed MinQ for key frames and alt ref frames relative to the ambient allowed range compared to what is inherited from VP9 and I have done some experiments in this regard.

There are two reasons as I see it for this.

1) AV1 using the same number of bits for a key frame is likely to be able to achieve a significantly lower Q due to its much more extensive intra coding options and the current min Q limits may be preventing us from spending the allocated bits.

2) AV1 may also, thanks to more and better inter coding options, may be better able to propagate improved distortion in key frames and ARFs to other frames which may help justify a bigger Q delta.

On balance though I would rather we seek to adjust this by changes to the boost calculation and the underlying KF minQ calculation (e.g. in get_kf_active_quality()) than through a separate constant factor applied here.

Also this will probably need to be considered alongside changes to the boost calculation.

Patch-set: 3
Reviewer: Gerrit User 5610 <5610@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>
Label: Code-Review=-1