Update patch set 2
Patch Set 2: -Code-Review

> Patch Set 2:
> 
> Why the difference between Q and VBR in this context?

In Q mode the deltas for Q on KF and ARF frames are not dependent in the same way on the bits allocated to the KF/ARF group. In VBR mode (and CQ mode) not coding a key frame at the scene change (from blank grey) a couple of frames after the start, on clips like Mob_cal, hurts the quality because there are not enough bits to do a really good job coding the first ARF in the absence of that kf at the real scene cut.

It would appear that the original patch helped psnr on mob_cal in Q mode but hurt a lot in VBR mode.

Patch-set: 2
Reviewer: Gerrit User 5610 <5610@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>
Label: Code-Review=0