Allocate mbmi on BLOCK_8X8 level for 4K+ videos
This commit change the allocation of cm->mi to one for each BLOCK_8X8
when the resolution is 4K or above.

Since most of the codebase assumes that the allocation done for each
BLOCK_4X4, we do not change how cm->mi_grid_base is allocated. i.e.
mi_grid_base is still allocated for each BLOCK_4X4, and some of them
are aliased to the same mi.

Total Memory Reduction:
2.6 GB => 2.5 GB, or ~4% reduction.

BUG=aomedia:2453

Change-Id: Ib3aa59761afffd063e1cff57becaf073ff7daff1