Update patch set 3
Patch Set 3: Code-Review+2

> Patch Set 2: Code-Review+1
> 
> (1 comment)
> 
> Does this CL change the behavior of rate control?
> Please let Paul approve this change.

I have cherry picked and run this on a few clips locally and my observation would tend to suggest that it results in more bits spent on the middle ARF than before, at least for easier clips like Akiyo and SignIrene. However, the reverse seemed to be true on hard content such as Husky.

In regard to Cheng's concern re rate control, I do not have data to be sure either way yet but feel that this change is more about distribution of bits between the ARFs at different levels  and that even if it does have an some impact on VBR rate control accuracy that it not a great concern at this point, but rather something to deal with later once we have the balance between normal and ARF frames right and variable sizing of ARF groups sorted.

Patch-set: 3
Reviewer: Gerrit User 5610 <5610@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>
Label: Code-Review=+2