Remove ediv_size_correction.
Removes the frame size correction to the error divisor used when
estimating the active worst quality in 2-pass vbr. In fact in VP9 the
adjustment based on frame size went the other way, but this could
be part of an interaction with other changes in VP9 not yet ported
over.

Change the default ERROR_DIVISOR to 96 such that the change is
neutral for 1080P.

Gives clear metric gains for low res and modest net gains for the
720p clips in the hdres set. The gain is more mixed for midres with
slight better ssim but down a bit on psnr-hvs.

STATS_CHANGED

Change-Id: I5053861baf0c9a215865485f1c47884fe64b5727