Remove ediv_size_correction.
Removes the frame size correction to the error divisor used when
estimating the active worst quality in 2-pass vbr. In fact in VP9 the
adjustment based on frame size went the other way, but this could
be part of an interaction with other changes in VP9 not yet ported
over.

Change the default ERROR_DIVISOR to 96 such that the change is
neutral for 1080P.

Substantial metric gains for lowres and ugc360p.
Results more mixed for midres.
? Small net gains for hdres (720p) but not all test points completed.

lowres  -0.597	-0.637	-0.866	-0.640
midres  -0.027	-0.031	-0.259	 0.183
ugc360p -0.785	-0.599	-1.435	-0.581
hdres   -0.077	-0.132	 0.009	-0.106

STATS_CHANGED

Change-Id: I5053861baf0c9a215865485f1c47884fe64b5727