Enable 3-way partitions for 16x16 blocks.
Only the 3rd block in this case is a chroma reference, and corresponding
chroma block will be of size 8x8. Most of the patch is to handle this
special case.

A few assembly functions were disabled due to alignment assumptions that
are no longer true.

Gain at speed 1 with CONFIG_3WAY_PARTITIONS on, compared to baseline
with CONFIG_3WAY_PARTITIONS off (ml_prune_4_partition speed feature
disabled for both for fair comparison):

VBR:
- lowres:  -0.111
- midres:  -0.169
- ugc360p: -0.124

Constant quality:
- lowres:  -0.094
- midres:  -0.081
- ugc360p: -0.094

Change-Id: I1b4b78ec551a7934926dee564c00225e7f0caacb