Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> (1 comment)
> 
> > Patch Set 3:
> > 
> > (1 comment)
> 
> I still see warning msg with that declaration, I would think we need to cast left and right of colon to specify both possible values?

this works for me:
return static_cast<uint8_t>((r < 128) ? r << 4 : r >> 4);

the warning is about the conversion done to the return type, not the shift on r (which is done as an int). So the outer cast is sufficient to cover the conversion.

Patch-set: 3
Reviewer: Gerrit User 5020 <5020@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>