Reduce decision threshold to use high precision MV
There is a slight improvement of the order of -0.03%
in BDRATE. Besides at medium to low quality there is a
speed improvement of the order of 5-10%.

Also added a speed feature to reduce usage of high prec
motion vectors even further, as well as to disable
high precision MVs completely.

The speed feature can be added to various speed settings
subsequently, especially the higher speed settings.

Change-Id: Idf0b1bf4e46058a25aedb3359ea914a7fa19ab0e