Use pyramid level rate correction factor.
Rate correction factor is used to estimate the required bits to
encode a frame.

The factor is determined by the factor of its nearest frame of the
same frame type. For example, if the frame is an internal arf, its
factor relies on the previous internal arf's factor.

Here we compute the rate correction factor for each layer separately,
and use the averaged value as the base for current frame's factor.

Problem:
This change works for two pass, but need some more work to make
it harmony with one pass.

Performance:

Change-Id: Ibd197d9dd907fed162126f7af7c3ee6197a8e185