Adjustments to KF Boost calculation.
Adjust the KF boost calculation to better match the calculation in VP9.
This monitors the difference between the last frame and second last
frame error and uses this to estimate the rate of decay in prediction
quality and ultimately trigger a break out in the boost calculation.

There are significant net benefits for all sets apart from the midres set
where the results are mixed (better vmaf worse ssim). In this set all
the clips are scaled from larger sources and this may be a factor. This
may point to a fruitful area for further changes.

In general there is a decline in boost for very easy clips and the worst
regression case is midres Snow_Mnt_480P which is almost static. In
contrast boost is increased a little for some harder clips like Husky and
ParkJoy.

Results:- (Overall PSNR, SSIM, PSNR-HVS, VMAF) -ve better.

Low Res -0.923	-0.119	-0.974	-1.878
UGC 360 -0.871	-0.608	-0.746	-1.203
Mid Res  0.080	 0.800	-0.133	-0.678
Hd Res  -0.514	-0.423	-0.579	-0.479

STATS_CHANGED

Change-Id: I6d1239bb6fb6fe7cd5e40a85787022edb03c738f