Update patch set 6
Patch Set 6:

(3 comments)

Sorry for the slow review responses: I've been out of the office.

I'm very concerned about how this API is supposed to be used by application, and about how the internals will be implemented within the library.

Primarily I don't believe this API should be HDR10+ specific. There are a couple of critical underlying issues:

1) The encoder does not need to know that the metadata is HDR10+, and it probably should not know that. The encoder only needs to know that the data is metadata. Beyond that the data should be opaque to the encoder. It only needs to know that it's ITU-T T.35 metadata.

2) The decoder has no way of knowing what a ITU-T T.35 metadata OBU carries in its payload: the AV1 specification does not define the format of ITU-T T.35 payloads. The decoder cannot know what the data is-- it knows only that ITU-T T.35 metadata has been encountered, and how many bytes of it are present.

Instead of adding hdr10p fields to the aom_image a pointer to a new struct should be added. Something like this should be sufficient:

typedef struct aom_metadata {
  OBU_METADATA_TYPE type;
  uint8_t *metadata_buffer;
  size_t metadata_size;
} aom_metadata_t;

And then a pointer to the struct and a count should be added to aom_image_t:

aom_metadata_t *metadata;
size_t num_metadata;

A simple data structure like a list might be better, but the above would work-- the drawback of this method is that the entire blob of metadata objects would need to be reallocated and copied each time a new metadata OBU was encountered in the decoder. The same would be true on the encoder side when the encoding application was inserting multiple chunks of metadata.

Note that I still have some memory handling related concerns about this patch, but the structural changes described above are of much greater import, and much of what you have here will likely be discarded. I've left my review comments for reference, but please focus on the design and spec issues pointed out above.

I've added to the CC line of this patch given that this is a relatively major library API change. Libaom, at present, simply discards metadata. A bug to track the required work here should be created as I believe that properly implementing this support will take a minimum of 3-4 patches, and probably a little more to get everything right. A series of patches like this would make sense to me:

1) Add aom_metadata_t and the changes needed to move it around and access it within libaom, and some basic tests to confirm that alloc and dealloc work as intended.
2) Add encoder support.
3) Add decoder support.
4) Add an end-to-end test that makes sure that it can be encoded by an encoder application and accessed by a decoder application.

Since this is an API change and libaom has downstream users we need to ensure that API level changes are something we can live with for the long term.

Patch-set: 6
CC: Gerrit User 9545 <9545@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>
CC: Gerrit User 5020 <5020@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>
CC: Gerrit User 5015 <5015@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>
Reviewer: Gerrit User 5215 <5215@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>