Update patch set 2
Patch Set 2:

Yes, good catch, this is redundant and shouldn't have gone into the spec.  The reason why the code came out this way was that the tap weights were fixed during the final tuning of CDEF and various weights for both the primary and secondary taps were tried, both based on the primary strength since the primary strength offers more freedom (16 values rather than just 4 for the secondary strength).  Gains beyond the noise level were only seen for changing the primary taps, but the code was never simplified, either by negligence or I possibly left it as it was to allow further experimentation at some point without adding the redundant code back.  Removing the code is no problem, and I doubt there will be retuning for AV2 at this level anyway.

I don't think the change requires an update in the spec change, but if the spec text is going to get updated anyway with other non-normative changes for redundancies, increased clarity, typo fixes and such, that could be an opportunity to change the cdef section as well.

Patch-set: 2
Reviewer: Gerrit User 5111 <5111@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>