Update patch set 8
Patch Set 8:

> Patch Set 8:
> 
> (1 comment)

Ruiling, yes, your are right, sse2 handles sad16xN better, but you are missing it is not the separate patch but a step to sad4d optimization. The functions you mention will process not 16 but 32 elements actually in the final patch. The corresponding changes are coming (Struggling hard to upgrade my experimental code to production quality and to split it into several commits). But right you are - it is not good to accept this change as is. I will fix it accordingly. THANKS!

Patch-set: 8