Update patch set 15
Patch Set 15:

> Patch Set 15:
> 
> This change is a bit large and looks to strip old behavior and add new behavior. How about removing the old behavior in a separate change list then adding the new functionality?

Hi, James

Just a bit history FYI, we had this proposal long time ago to add the config file for passing control flags. The initial patch was just a place holder, but it was never fully implemented and used. This patch clean it up and make it fully functional.

Patch-set: 15